+++ src/main/java/roomescape/reservation/controller/ReservationAdminController.java
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.reservation.dto.ReservationResponse;
-import roomescape.reservation.dto.ReservationSaveDto;
+import roomescape.reservation.dto.response.ReservationResponse;
+import roomescape.reservation.dto.request.ReservationSaveDto;
 import roomescape.reservation.service.ReservationService;
 
 @RestController
@@ -34,7 +34,7 @@ public class ReservationAdminController {
 
     //TODO: status 변경으로 수정
     @DeleteMapping("/reservations/{id}")
-    public ResponseEntity<ReservationResponse> delete(@PathVariable Long id) {
+    public ResponseEntity<ReservationResponse> delete(@PathVariable Long id ) {
         return ResponseEntity.ok(reservationService.delete(id));
     }
 }
