+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -14,8 +14,8 @@ import roomescape.common.exception.NotFoundException;
 import roomescape.date.domain.ReservationDate;
 import roomescape.date.fixture.FakeReservationDateRepository;
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservation.dto.ReservationResponse;
-import roomescape.reservation.dto.ReservationSaveDto;
+import roomescape.reservation.dto.response.ReservationResponse;
+import roomescape.reservation.dto.request.ReservationSaveDto;
 import roomescape.reservation.repository.FakeReservationRepository;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.fixture.FakeThemeRepository;
